Delegate Entry messages to Entry.update
Having trouble getting this to work because when an entry is opened for
editing we need to return a Cmd to switch the focus to the input.

I'm also wondering whether we need to seperate the Entry datatype from
the component which represents an entry in a list. E.g. should Entry
store whether it is being edited?